OS11SPRYKE-1714 product store relation - jenkins job [5.9] (#2215)
* OS11SPRYKE-1714 product store relation

* Product store relation

* OS11SPRYKE-1714 new product store relation

* OS11SPRYKE-1714 new product store relation

* OS11SPRYKE-1714 new product store relation

* OS11SPRYKE-1714 new product store relation

* OS11SPRYKE-1714 add ff event executor

* OS11SPRYKE-1714 add ff event executor

+++ src/Pyz/Zed/FactFinderExport/Business/Model/FactFinderEventManager.php
@@ -60,8 +60,10 @@ class FactFinderEventManager
      */
     public function clearEvents(string $eventName, array $idProductAbstractList)
     {
-        $sql = "delete from pyz_fact_finder_event where event_name = '" . $eventName . "' and fk_product_abstract in(" . implode(",", $idProductAbstractList) . ")";
-        $this->getResult($sql);
+        if(count($idProductAbstractList) > 0) {
+            $sql = "delete from pyz_fact_finder_event where event_name = '" . $eventName . "' and fk_product_abstract in(" . implode(",", $idProductAbstractList) . ")";
+            $this->getResult($sql);
+        }
     }
 
     /**
